Extract shared AI-digest helpers into ai-digest-utils
The article-digest and law-summary services each carried byte-identical
copies of the text/cache/single-flight plumbing. Pull the genuinely shared
helpers (stripTags, clip, normalizeText, stableHash, single-flight, cache
load/save, JSON extraction, citation grounding) into shared/ai-digest-utils.js
and import them.

loadCache/saveCache now take the cache filename, and normalizeCites takes a
{ limit }; per-service specifics stay local (article's 3-key cacheKey, the
summary's sentence-boundary clip and its own extractJsonObject message).
