+++ backend/shared/article-digest-service.js
@@ -1,8 +1,14 @@
-const fs = require('fs');
-const path = require('path');
-const crypto = require('crypto');
-
 const { chatComplete } = require('./openrouter-chat');
+const {
+  clip,
+  normalizeText,
+  stableHash,
+  makeSingleFlight,
+  loadCache,
+  saveCache,
+  extractJsonObject,
+  normalizeCites,
+} = require('./ai-digest-utils');
 
 const CACHE_FILE = 'article-digest-cache-v1.json';
 const CACHE_VERSION = 1;
@@ -12,7 +18,7 @@ const CASE_LAW_CACHE_VERSION = 'case-law-cache-v4';
 const MAX_ARTICLE_TEXT_CHARS = 5500;
 const MAX_DECLARATION_CHARS = 1800;
 
-const inFlight = new Map();
+const withSingleFlight = makeSingleFlight();
 
 const SYSTEM_PROMPT = `You write concise digests of how CJEU case law interprets one article of an EU legal act.
 
@@ -36,62 +42,12 @@ Rules:
 - Prefer 2-5 themes.
 - If the input contains no matching cases, return {"summary":"","themes":[],"noCaseLaw":true}.`;
 
-function stripTags(value) {
-  return String(value || '')
-    .replace(/<[^>]+>/g, ' ')
-    .replace(/&nbsp;|&#160;/g, ' ')
-    .replace(/&amp;/g, '&')
-    .replace(/&lt;/g, '<')
-    .replace(/&gt;/g, '>')
-    .replace(/&quot;/g, '"')
-    .replace(/\s+/g, ' ')
-    .trim();
-}
-
-function clip(value, maxChars) {
-  const text = stripTags(value);
-  if (text.length <= maxChars) return text;
-  return `${text.slice(0, maxChars).trim()}...`;
-}
-
 function cacheKey(celex, articleNumber, lang) {
   return [
     String(celex || '').toUpperCase(),
     String(articleNumber || '').trim(),
     String(lang || 'ENG').toUpperCase(),
   ].join('_');
-}
-
-function stableHash(value) {
-  return crypto.createHash('sha256').update(JSON.stringify(value)).digest('hex');
-}
-
-function withSingleFlight(key, factory) {
-  if (inFlight.has(key)) return inFlight.get(key);
-  const promise = Promise.resolve()
-    .then(factory)
-    .finally(() => inFlight.delete(key));
-  inFlight.set(key, promise);
-  return promise;
-}
-
-function loadCache(cacheDir) {
-  try {
-    const filePath = path.join(cacheDir, CACHE_FILE);
-    if (!fs.existsSync(filePath)) return {};
-    const parsed = JSON.parse(fs.readFileSync(filePath, 'utf8'));
-    return parsed && typeof parsed === 'object' ? parsed : {};
-  } catch {
-    return {};
-  }
-}
-
-function saveCache(cacheDir, cache) {
-  fs.mkdirSync(cacheDir, { recursive: true });
-  const filePath = path.join(cacheDir, CACHE_FILE);
-  const tempPath = `${filePath}.${process.pid}.${Date.now()}.tmp`;
-  fs.writeFileSync(tempPath, JSON.stringify(cache, null, 2), 'utf8');
-  fs.renameSync(tempPath, filePath);
 }
 
 function matchesArticle(c, celex, articleNumber) {
@@ -151,64 +107,6 @@ function buildArticleDigestInput(celex, articleNumber, parsedLaw, caseLawPayload
     } : null,
     cases: matchingCases,
   };
-}
-
-function extractJsonObject(text) {
-  const trimmed = String(text || '').trim()
-    .replace(/^```(?:json)?\s*/i, '')
-    .replace(/```\s*$/i, '')
-    .trim();
-  const match = trimmed.match(/\{[\s\S]*\}/);
-  if (!match) {
-    const snippet = trimmed.replace(/\s+/g, ' ').slice(0, 500);
-    throw new Error(`Digest model did not return a JSON object; text=${snippet || '<empty>'}`);
-  }
-  return JSON.parse(match[0]);
-}
-
-function normalizeText(value, maxChars = 1200) {
-  const text = stripTags(value);
-  return text.length > maxChars ? `${text.slice(0, maxChars).trim()}...` : text;
-}
-
-function buildCitationIndex(input) {
-  const byEcli = new Map();
-  const byCelex = new Map();
-  for (const c of input.cases || []) {
-    const declarationNumbers = new Set((c.declarations || []).map((d) => String(d.number)));
-    if (c.ecli) byEcli.set(c.ecli, { ...c, declarationNumbers });
-    if (c.celex) byCelex.set(c.celex, { ...c, declarationNumbers });
-  }
-  return { byEcli, byCelex };
-}
-
-function normalizeCites(value, input) {
-  const { byEcli, byCelex } = buildCitationIndex(input);
-  const seen = new Set();
-  return (Array.isArray(value) ? value : [])
-    .map((cite) => {
-      if (!cite || typeof cite !== 'object') return null;
-      const ecli = normalizeText(cite.ecli, 80);
-      const celex = normalizeText(cite.celex, 40);
-      const declarationNumber = normalizeText(cite.declarationNumber || cite.declaration || cite.paragraph, 20);
-      const source = (ecli && byEcli.get(ecli)) || (celex && byCelex.get(celex));
-      if (!source) return null;
-      const normalizedDeclaration = source.declarationNumbers.has(String(declarationNumber))
-        ? String(declarationNumber)
-        : null;
-      const key = `${source.ecli || source.celex}:${normalizedDeclaration || ''}`;
-      if (seen.has(key)) return null;
-      seen.add(key);
-      return {
-        ecli: source.ecli || null,
-        celex: source.celex,
-        caseNumber: source.caseNumber || null,
-        name: source.name || null,
-        declarationNumber: normalizedDeclaration,
-      };
-    })
-    .filter(Boolean)
-    .slice(0, 6);
 }
 
 function parseArticleDigestJson(text, input) {
@@ -307,7 +205,7 @@ async function ensureArticleDigest({
   const key = cacheKey(celex, articleNumber, lang || input.lang);
 
   return withSingleFlight(`article-digest:${key}:${sourceHash}:${model}`, async () => {
-    const cache = cacheDir ? loadCache(cacheDir) : {};
+    const cache = cacheDir ? loadCache(cacheDir, CACHE_FILE) : {};
     const cached = cache[key];
     if (
       cached?.version === CACHE_VERSION
@@ -339,7 +237,7 @@ async function ensureArticleDigest({
         generatedAt: new Date().toISOString(),
         digest: generated.digest,
       };
-      saveCache(cacheDir, cache);
+      saveCache(cacheDir, CACHE_FILE, cache);
     }
 
     return {
